LayoutTests/imported/w3c/web-platform-tests/dom/nodes/moveBefore/moveBefore-shadow-root-expected.txt


PASS moveBefore() is allowed in ShadowRoots (i.e., connected DocumentFragments)
